Update 10/08/2021 – Line Of Sight Bugfixing 2: The Pathfinding
I fixed some more AI pathfinding bugs.

For starters, a common problem is that the AI would break line of sight with the player, but wouldn’t find a new position. After checking, I found out that I did not invert a bool. The AI is supposed to find a new position if it can’t find line of sight, but it was finding a new position if it could. I fixed this, and the AI started rapidly switching positions, covering the entire screen with debug lines as it rapidly searched for new locations. After checking what the line of sight check was hitting, I found that it was catching on the player collider itself, a similar problem to what occurred with my AI aiming code. Like with that problem, I needed to solve it with a list of both the player and enemy colliders, but I had already created such a list in the AIAttack class. To solve this, I made a property in the AIEntity class that calculates such a list.

The enemy now appears to be correctly determining what is and isn’t in its line of sight, but there are still problems to fix. Most notably, when searching for a new position, the enemy will often get stuck in one place, as the debug lines I drew for pathfinding jitter quickly between two separate locations. Theorising that the code was returning positions that were null or which the enemy could not appropriately reach I wrote some debug logs to figure out which stages of the checking process were causing so many wonky results. I found another instance of line of sight code that didn’t properly account for the AI and target colliders, and fixed it, but the problem is still persisting.