Update 13/08/2021 – Slogging through debugging pathfinding code, also experiments with overhauling EventHandler
I did some more work on solving the AI issue where the enemy would occasionally rapidly switch between two destinations, unable to move due to continually resetting its movement path. This turned out to be a much greater can of worms than I’d anticipated. After several hours of fruitless debugging involving breaking the code to find each reference to changing the destination, figuring out that NavMeshAgent.SetDestination is a bool and not a void, removing the NavMeshAgent during runtime to study the resulting MissingReferenceExceptions when other scripts tried to access it, liberal misuse of the Find References in Scene tool, and an ungodly amount of precisely placed print messages and debug lines, I was unable to find the solution. I did, however, notice some recurring traits when it came to the glitches.

First, the new position would generally be higher in altitude than the intended position. Second, the new position is assigned on some kind of timer, as it is repeatedly assigned at a constant rate before being reverted back to the main point and does not get reapplied when the game pauses. And last, whenever the flickering between positions occurs, the desired position always seems to be inside a wall, in a point where there is technically a valid NavMesh surface but no way for the agent to get to it.

This last point seems to be the most relevant. The EngageTarget code that the AI uses to find an appropriate target is supposed to have checks to eliminate this sort of thing. The function for finding a new position consists of a multi-stage checking process. This process is repeated several times in a for loop, to find the best position out of several samples.
1.	A random position is sampled around the player, within a certain set of range requirements so it is not too close or too far.
2.	NavMesh.SamplePosition is used to check if there is a position close by that is guaranteed to be on the NavMesh.
3.	A line of sight check is performed to ensure the AI can see and attack the player while standing at said position.
4.	NavMesh.CalculatePath is used to ensure the position can actually be reached by the agent.
5.	If this is the first check, the path is stored for the duration of the loop. For each subsequent check in the loop, the newly formed path is checked to see if it is shorter and therefore easier to get to than the current shortest path.
After the loop is completed, the remaining shortest path will be assigned as the new destination for the NavMeshAgent.

I think the culprits might be either NavMesh.SamplePosition or NavMesh.CalculatePath. I am not completely familiar with how to use these functions correctly, and it is possible I am using them wrong and in a way that is returning dodgy positions, causing other glitches. After some googling (but at a point too late in the day for me to spend time testing in this update), I discovered that NavMesh.CalculatePath doesn’t necessarily guarantee a functioning path even if it returns true. It can return true but produce a partial path, if the agent is unable to reach its destination.

Fortunately, I can check the path’s status, which is stored as an enum. Why it would be relevant to know about an incomplete path is beyond me, as is why they would have the function return a bool and assign a separate enum value that determines the actual path status, when the function could have simply returned the enum. In any case, I will need to add another check for the path status enum, as well as more extensive testing of the NavMesh.SamplePosition function.

I also experimented with some new code for the EventHandler system. After discussing the setup with other programmers in a chat room, I realised that my system may be overly complicated, with the EventObserver element being somewhat redundant. I experimented with some new code. One variant involved putting the delegates that are in the EventObserver, directly into the EventHandler. I also made an easy to use static singleton for referencing the EventHandler, which requires absolutely no additional setup while also not relying on rapid calls of GetComponent, which can be bad for performance.

I also experimented with another variant involving an interface. My idea was that any MonoBehaviour object that would want to receive game events could use an event named IEventReceiver, with appropriate event response functions therein. When ‘transmitting’ a message, the function doing so could use the EventHandler to directly run the function by referencing all IEventReceivers in the scene, without the object with the IEventReceiver needing to add any functions to delegates. I’m not sure if this function will work, however, because the IEventReceiver itself still needs to be added to and removed from a central list.

I don’t know how to make an interface have functions that run automatically, meaning I would still have to add and remove the interfaces manually in the OnEnable and OnDisable functions of whatever MonoBehaviour script they are attached to, and it is much more difficult to search for interfaces in a scene, compared to MonoBehaviours. So I feel that in the long run, continuing to use delegates but without the EventObserver may be the better option. But again, it’s too late and I’m too tired to work on those things, so I’ll have to add them to the list.

I also quickly made a static class, Colours, to easily reference colours for debugging that weren't included as simple options in the default Color class. It wasn't very interesting or complicated to create, so it's in the footnote, but it's nice to have more colours for differentiating information.