Update 16/08/2021 – 17/08/2021: Improving GUI, gameplay options menu, attempts at getting proper key binds and a possibly improved input system
I continued some work I’d started a while back on nesting the GUI elements as prefabs, by making several prefab variants. As a first-person shooter is a much more complex game, there are a lot of GUI elements I need to manage, especially for options menus.

I started making a menu for gameplay settings, because I realised this is both one of the most essential and one of the easiest to implement. My options menu code is a bit of a mess, I may need to start from the top by creating a base abstract OptionsMenu class and making inherited versions for separate menus for all types of options to adjust. This may have been something I’ve already tried doing, but with the many projects and tasks vying for my attention it’s easy for things to slip down the priority list. However, as one of my main objectives with this project for the foreseeable future is to make this into a presentable demo, a functional options menu is a must, therefore it is something I will have to presently prioritise for this project.

I experimented with creating a timer that does not need to be continually incremented during Update(). I did this by, instead of storing a float for elapsed time that continually ticks up, having a float representing the real time the timer was reset. I can then check this against the current time and time scale, with only slightly more calculations at the time, but without needing to continuously monitor it during Update(). That being said, I realised that I was using it to procrastinate, and quickly shelved it to work on more important features, such as the options menu I listed beforehand.

Realising that a gameplay settings menu will also mean a button remapping menu, I had a look at my game’s input system. My game currently uses the old Unity Input system, which works well enough for everything currently implemented, but there are still several things I want to add, the most notable feature being the ability to bring up visual button prompts for various keys, mouse buttons or controller inputs. After experimenting, the best method I can think of with the current input system would be to get a KeyCode enum value of the key, then scroll through a massive switch statement to find the matching value, and pull an icon from a corresponding file path string. The daunting effort required for this, combined with the knowledge of Unity’s new and seemingly improved input system, meant I decided to work on the other options menu elements instead.