Update 19/08/2021 – Testing gameplay options menu, creating graphics options menu
Finished up and tested the brand new GameplaySettingsMenu. It works! Mostly. One bug I noticed is that upon resuming the game, the player’s FOV will only change if they are holding a weapon with a scope function. There is also an undesirable cosmetic effect where the apply and revert buttons will appear completely whited out if they are disabled. This should be easy to fix without altering any code, though.

Created a GraphicsSettingsMenu. The menu currently has four settings – fullscreen mode, graphics quality preset, resolution and framerate. This was slightly more complicated than the previous menu, primarily due to the resolution and framerate settings. This is because in Unity, when asking for a list of all available screen resolutions, it delivers a single-dimensional array of Resolution structs, each containing a different resolution and refresh rate. This leads to the problem where the player goes to select a resolution, and is given 5-6 versions of each resolution, because there are, say, six different versions of the 1600x900 resolution but with varying frame rates. But I couldn’t just remove all values of a particular resolution after I record it for the first time, because I want the player to be able to record their frame rates as well. But I couldn’t just record a list of every possible resolution plus a list of every possible frame rate, because some might not be compatible.

Fortunately, I researched and found two things that could help me. The first was a 2-dimensional array, where a length and height are specified and it creates a grid, where appropriate items can be selected using two indexes as grid coordinates. This was promising, but would lead to a lot of unnecessary empty spaces where a resolution does not support a particular frame rate or vice versa, and since Unity’s Resolution variable is a struct, I cannot null them. Fortunately, I found out that it is completely possible to create an array consisting of more arrays, each of which I can give a unique length. I then created a static property that checks the total number of different resolutions, and the different available frame rates for each solution, and generates an array of arrays. Each sub-array contains every available frame rate for a given resolution, with the values being sorted by resolution first since it has a more immediate impact on gameplay, and therefore players will want to prioritise setting it first. I was then able to use the now-sorted variables to generate a dropdown list of all available resolutions, that upon selection will automatically update a second dropdown to show all available frame rates for said resolution. Upon applying changes, my code will use the selection values of the two dropdowns as indexes to find the appropriate Resolution struct from the table, and apply it.

I’m very pleased with this code, except for the fact that I haven’t tested it yet. These two menus are lacking in many features that would need to be implemented down the line. There are many options I still need to implement, and the ‘revert to default’ function currently does nothing as I don’t yet know a way to get ‘default’ graphical settings. Overall, however, I am pleased with the core framework and progress I’ve made.