Update 19/08/2021 – Debugging options menu, fixing ammunition type validation
I ‘completed’ and implemented my video options menu, and started testing it. Despite my proud efforts, I’m still faced with a lot of problems to fix. After adding functions to refresh the dropdown options, so the placeholder options did not appear, I was faced with the problem of my resolution and refresh rate dropdowns not showing any options.

After debugging with print messages, I confirmed that the length of the first column was set to zero, and no values were being recorded. I thought, since the first Resolution struct of each column was assigned manually, that I needed to manually set the appropriate column length variable to 1 as well, and that failing to do so caused some sort of mismatch. I fixed this, and the resolutions started displaying fine, but there was only one refresh rate option for each resolution. After checking debug logs, I found that for some reason, none of the column lengths were being increased past 1, and therefore only one refresh rate was being stored, which was curiously always the maximum specified refresh rate for the monitor (I was using a 240Hz monitor and altered the value between that and a more conservative 60Hz).

I created another string that listed all Resolution values in Screen.resolutions, without any sorting, and found out that the system was working as intended. The property wasn’t getting any different refresh rates because Screen.resolutions was only including Resolution variables with the same framerate.After discussing it with someone in a chat room, we theorised this was because I was running it in Editor mode. After quickly making a build and testing it out, I was able to get the refresh rates to appear properly. But I noticed a new problem where besides the fullscreen mode, I had tremendous difficulty getting the settings to save properly. This is something I will need to solve later.

I noticed that I was unable to navigate the interface using keyboard buttons, only using the mouse. I realised that the EventSystem and Standalone Input Module, two components required for that to work, were not actually attached to the pause menu, to which the options menu is presently attached (although it’s fortunately a prefab that I can easily drag into any other menu). I copied them onto the Pause Menu prefab and applied them as permanent changes to the prefab. I then also added code to, every time the player switched to a new window while in the pause menu, to automatically update the first selected game object appropriately. This, however, only partially worked, and the player can navigate the initial pause menu using the keyboard, but it will not switch to a new button if the player switches to a new menu.

I fixed a recurring problem with the player’s ammo inventory system. It consists of a list of different Resource variables (a custom set of variables containing max, current and critical ints, and an identifying name string) for the various ammo types. Instead of adjusting the amount of types in the array, it is meant to update automatically, populating itself based on an AmmunitionType enum. This is intended to allow myself (or other developers) to easily add and remove ammo types, and select the appropriate one from a dropdown list. The old system was something I copied off the Internet a while back, whose code I didn’t understand, which eventually broke and stopped updating properly, essentially rendering me unable to update available ammunition types. I eventually bit the bullet and remade it from scratch, and I can say with confidence that it works perfectly.

When validating the array of types, it converts the array into a list, for sorting purposes. The validation function uses System.Enum.GetNames to obtain an array of name strings for all ammunition types. Since this also indicates how many types there are, a new Resource array is created with an appropriate length. A for loop is used to update each Resource in the array, by searching in the old list for a value with the same name. This is essential, as otherwise all implemented stat values for the ammunition inventory would be wiped any time any value was changed. The old resource is then copied into the new array at the appropriate position – if one is not available, a new one is created instead. Once all appropriate Resources are copied or created, the old array is overwritten by the new one. Any Resources for deleted ammo types are left ignored in the old array, and overwritten.

I didn’t get quite as much work done today as I normally would have, since I was busy writing a CV and attempting to recover from staying up until 4PM last night. But I’m still pleased with the work I was able to get done.