Update 20/08/2021 – 21/08/2021: Attempting to fix AI dodging and line of sight checks, also optimising health, exploding barrels and several very minor changes
Attempted to fix bug where AI would not dodge attacks

The first presumed reason was that the attack messages containing information about attacks would not actually register anything as being in the line of sight.

After searching in my code, I found a possible cause. The complex line of sight checking function I am using relies on generating a grid of raycasts to sweep the area in front of each collider, but rejecting directions that fall outside the correct angle. I found that at certain distances, the length of spacing between raycasts in the grid was actually slightly larger than the area covered by the attack’s cone of fire. This would often mean that no valid raycasts would actually land inside the cone of fire, and therefore nothing could be checked. I added code detecting the diagonal distance between raycasts (i.e. the maximum possible distance), and checking if it was smaller than the cone of fire. If so, the raycast spacing distance is reduced to the cone diameter, with a multiplier of 0.75 to ensure it is always slightly smaller and that the cone will always have at least one raycast inside of it. While this was a legitimate issue that needed fixing, it does not seem to have solved the original problem.

I added a ‘time initiated’ variable to my attack messages, so enemies can anticipate how quickly they will need to dodge them. Admittedly, this variable is not actually referenced in any of the code, yet.

Started typing up a prototype script for an enemy with pathfinding that is capable of flight. I did not do much work, as I realised I had other more important things to focus on.

Added some new values to the DamageType enum. This enum currently does very little in its current state, and I will need to overhaul it at some point. I anticipate this will be easier to do, but I have more important game elements to create first.

I created a dedicated ‘Explosion’ MonoBehaviour, that I intend to use with all game elements that produce explosions. This class contains all the necessary stats for an explosion, such as damage, knockback, blast radius and falloff, as well as auto-triggered events for cosmetics.

I attempted to create a simple exploding barrel, but it did not react when tested. I believe this may be due to its Entity script not possessing a faction variable, and my damage code thinking it may be an ally. This is something I will need to look into later, but which is relatively low-priority.

I altered the Health script to not utilise Update(), saving processing power during runtime. The only code in it was an if statement checking if the entity’s health was above zero, and activating the death sequence if so. I added this code to the public function for taking damage, since it only needs to be checked when their health value changes.