Update 24/08/2021 - Finally solved the dodging problem!
The dodging problem was a side effect of transitioning to the new enemy that has an actual 3D model and animations (even if they’re placeholders I downloaded for free off the Unity Asset Store). Since I had to throw out the old model and collider and replace them with a new model and set of colliders, that were on a different layer to the rest of the entity. Since the line of sight checking code I used is very processor-intensive when applied to everything in a scene, I added a secondary check involving Physics.OverlapSphere and a second LayerMask, to only scan for relevant objects in the scene (i.e. ones on the ‘Entity’ layer). Since the enemy’s colliders were not on this layer, the attack risk checks would ignore the enemy colliders and as a result the game would not register any of the enemies as being at risk of being hit by attacks. After fixing this, the enemies did proceed to dodge attacks! Hooray!

That being said, I’ve decided the specific dodging action is awkward, unrealistic and unfun, with the enemy darting around like a cartoon ninja and being annoyingly difficult aim at with most weapons. While the current behaviour would be appropriate for a single enemy type, it would be very annoying for most of the enemies, since it means the shotgun is currently the only weapon that feels really satisfying to use. But that’s something I need to fix next time. I didn’t do much work on this project today because I had other projects to work on.