Update 06/09/2021: Implementing ragdoll physics, optimising horrible lag spikes, weapon holding positions
Fully created ragdoll for placeholder soldier model! Hooray! The joints seem to be a bit too loose, with certain joints on the model bending in ways they shouldn’t realistically be able to, but it works well enough that I have more important things to focus on.
Attempted to implement ragdoll to my enemies so it would actually be used in gameplay.
I created a Die() function in the enemy’s AI script, that disables all their behaviours and other required components. I then added that to the OnDeath() event in the enemy’s health script, as well as functions for disabling their animation handler and enabling the ragdoll. Now, instead of disappearing upon death, the enemy will flop to the ground as a corpse. This is a massive improvement for realism and presentation, whereas formerly the enemies would simply disappear when killed.

However, it has been a while since I attempted to properly playtest my levels, and while testing the ragdoll functionality I have become highly aware of drastic performance dips while playing, even on my reasonably powerful computer. After learning some tips on using the Profiler from some helpful people in a Discord chat room, I found the code was predictably coming from my AI script’s Update() function, specifically from the garbage collector. A lot of data was being produced, clogging up memory, meaning that a lot of processing time was spent removing it each frame.
Since that much garbage will exist as a result of things being created and abandoned, I tried looking through the function to find all instances of new data instances being created.
One thing I noticed was the property for reference and combine the hitboxes of the enemy and target. I realised this was causing a massive amount of superfluous data, as every time it was called it was creating a new array and using for loops to run through functions to populate it. While I’m not entirely sure this was the cause of the performance slowdown, I now knew a good place to start, by optimising the function to work in a more optimised manner.

I also did a small amount of work on having actual weapon models for guns. I was worrying earlier about how to have unique holding positions for each weapon, when I wanted the model and collider to be where the actual transform axis is so when they’re dropped they can be moved about easily. But I realised that I can still do that by having the separate transform exist as a data reference, but having the weapon model simply switch between its origin and the holding position when dropped/picked up. I will need to eventually bite the bullet and actually test these.