Update 07/09/2021: Debating optimisation solutions for AI line of sight
I started work on attempting to fix the line of sight code and optimise the combined array for agent and target hitboxes, which according to the Profiler were the main sources of lag and increased CPU usage.
After wondering over many ways to optimise the combined array, I’m considering it might be better to avoid using it. Since the combined hitbox array exists to ensure the line of sight check does not return false due to hitting a collider that’s actually part of itself or the thing it’s looking for, I realised it would be better to just use two separate arrays for the purpose. This not only saves the trouble of assembling an array, but it also might mean I can immediately return true if the raycast hits a collider that is part of the target, saving unnecessary calculations.
This may not work due to the RaycastAll function (necessary so if it hits something it’s supposed to ignore it can still accurately check something else in its path) not ordering its results in any particular way, but I may be able to fix this with a list-based sort function. The fact that the line of sight check relies on making a new array every time it’s called may also contribute to the garbage collection issue like the combined array does, but it’s a static function, I might have to make a static list that is continuously updated (although I don’t know if multiple functions running at the same time will interfere with each other’s raycast hits). Either way, I have a solid direction to work in in order to make my game run quicker and smoother.