Update 08/09/2021: Optimising line of sight code, partially fixing animations
Last time, I was considering having a static list of RaycastHits for my line of sight check, to ensure less garbage was being created. I did not go with that, but I did fully implement my plan to make the check work with two separate arrays. This is ironically the best solution I can think of, since by having it differentiate between target colliders and exceptions, it can prematurely end the function if true and avoid unnecessary processing power, meaning combining the two hitbox arrays together is actually disadvantageous.

I also fixed a problem with my placeholder enemy’s animations. The enemy would glide along the ground before starting their walking animation, and would often stay stuck in the walking pose even when standing still. I realised that many of the animation controller’s states had exit times, delaying the actual transition. I quickly fixed it and the enemy’s animations became much smoother. There is still the problem of the enemy’s walk animation moving forward even when moving backwards, which is something I tried to fix today, but the result somehow looked even worse.

I also did a bit of experimenting with a system to pre-calculate cover positions at the same time that the NavMesh is created, so during runtime enemies have less calculation to perform. It would involve launching a grid of RaycastAll functions down through the level to obtain positions on the floor, then check in a small radius for walls and store each position that is close to a wall in an array. Enemies would be able to look through the array for an appropriate cover position. I am not entirely committed to this, but I do want to do something to improve my enemy’s ability to seek targets and find cover. Currently, both these functions involve making a number of checks in a random position around the player or enemy’s position. I need to change this mechanic so it checks in a more orderly manner, perhaps by specifying a grid of positions to use, or slowly rotating and extending a Vector3 to arrange positions in a spiral pattern.

I also slightly changed the avoid attack function, to have a minimum distance as well as a maximum one. I felt this was important because with no minimum position, my enemies appear to dodge attacks far too confidently, moving a very short distance to dodge bullets whizzing right past their heads. This would be useful for some elite enemies, but I feel most enemy types would panic slightly more than that, and always try to put a lot more distance between themselves and the danger zone.