Update 08/09/2021, part 2: Fixing miscellaneous bugs that I noticed on a different device
Fixed the logic in the attack avoidance function, so if the agent cannot find a safe dodge position it aborts the function and resumes normal behaviour.
Fixed a bug where if the enemy stops engaging the player due to being out of range, a NullReferenceException can occur when checking if the player is dead.
Fixed a bug where my door animation had an unused event with improperly assigned parameters, causing an error when playing.