Update 09/09/2021: Experimented with more even methods for AI to find new positions
My current AI currently looks for new positions by performing a number of checks. For each check, it randomly selects a point within a certain radius, then accepts or denies it based on certain criteria. This is not a good system because of its inaccuracy – it’s possible for all the points to be clustered together and not notice an ideal position. I experimented with some algorithms for producing points in a roughly even space.

I vaguely toyed with a square grid, but I put more effort into a circular one. I experimented with a function that, using a position, rotation axis, minimum radius, maximum radius, number of loops and total number of checks, I could create a series of points moving out in a spiral pattern, for slightly more uniform distribution. I personally feel it’s actually a bit too uniform, without providing the full coverage of a square, but I’m still pleased with it. It needs more work though. I didn’t do as much work on this today because of other obligations, such as doing a practical test for a job application.