Update 29/09/2021: Adding some sound effects to the player, looking through old code, studying game feel effects
Finished implementing several player movement noises, for walking, jumping and landing, as well as drawing and holstering weapons. Although there are still a lot more aesthetic elements that need to be implemented, and I’m not 100% pleased with the specific sound files I currently have, these additions have drastically improved the game’s feel.
This required reworking some of my code, that I realised was half-finished, to ensure the sound effects worked properly. This led to an unpleasant realisation at how awful a lot of my old code was.
I also spent some time playing similar existing games and documenting various aesthetic effects, that aren’t obvious at first glance but which greatly improve their feel. This is essential to polishing the aesthetic presentation of this project to make it work better in a portfolio.