Update 30/09/2021: Cosmetic effects for player controls, and cleaning up redundant features
Implemented multiple simple cosmetic effects to greatly improve the feel of the player controls.
•	A visual effect that appears when the player crouches, for a clearer indication of their movement state. This effect consists of a dark gradient at the bottom of the screen, and a small icon of a person crouching.
•	Improved and simplified smooth lerping functionality for position and rotation of player torso/hands, allowing the animations to smoothly interpolate.
•	The upper body will tilt when the player moves around, emphasising the momentum of movement. This can be set to a positive value to make the player’s top half lean towards the movement direction, implying the player is throwing their weight into movement, or a negative value to have the torso flail about like a cartoon character.
•	When the player is standing still, a simple looping idle movement will play, simulating the player character breathing.
Moved and tweaked Heads Up Display function from the player object itself to the canvas where the actual HUD elements are, for a more modular design and less separate scripts cluttering up the Player object.
Removed the minimap as it was not serving a practical purpose, and did not look good enough to justify keeping around.
Removed inherited classes PlayerHealth and NpcHealth, as there was no functionality in them that could not be achieved through the use of events in the base Health class.
Experimented with separate time scales for individual game entities, allowing different characters to experience different rates of time. I do not presently plan on using said feature for this game, and had tried creating it as a proof of concept to explain to another person, who was trying to create a similar system, but I believe it may come in handy for other games.